core: fix a bug and make sure that mounts do not propagate to parent namespace
We are using variable 'n' to calculate how many mounts we may add to the
new namespace at same time that var is updated later when we have
duplicate, invalid or nop mount paths, in this case n can be 0 and we
may not remount / as SLAVE. While we are it split some code out.
Reported by Evgeny Vereshchagin.